Fix the cast issue
git-svn-id: https://edk2.svn.sourceforge.net/svnroot/edk2/trunk/edk2@3972 6f19259b-4bc3-4df7-8a09-765794883524

+++ MdeModulePkg/Core/Pei/Image/Image.c
@@ -403,7 +403,7 @@ Returns:
     AsciiString = PeCoffLoaderGetPdbPointer (Pe32Data);
     
     if (AsciiString != NULL) {
-      for (Index = AsciiStrLen (AsciiString) - 1; Index >= 0; Index --) {
+      for (Index = (INT32) AsciiStrLen (AsciiString) - 1; Index >= 0; Index --) {
         if (AsciiString[Index] == '\\') {
           break;
         }
